Fix overflow when defining more than 255 actions (possible under NEW-PARSER?).

+++ Zilf/ZModel/ZEnvironment.cs
@@ -114,7 +114,7 @@ namespace Zilf.ZModel
         [CanBeNull]
         public ZilAtom LowDirection;
 
-        public byte NextAction;         // V? (intentions)
+        public ushort NextAction;         // V? (intentions)
 
         public int HeaderExtensionWords;
